Improved Accounts Delete Behavoir:
If data of a certain patient is deleted all his medical records will be deleted too.
If someone tries to delete an account the following rules apply:
1) Patient: If the patients data is still in the system the delete wont be executed. There will be an error message saying that the patients data needs to be deleted first.
2) Doctor: If the account id of the doctor is still used in the patients data table it is not possible to delete the doctor. There will be an error messag saying that all patients need to be transfered to another doctor in order to delete this account.
3) Admin: if an admin account should be deleted, the will check if there is more than one admin account. It is not possible to delete the only admin account via REST Interface.

This commit also fixes the following issue:
Lets say a patient is created, the id of the account is N and the system issues an access token. Due to the fact that MySQL keeps the current auto increment value for the id in main memory, it is possible to assign this id to a new account after deletion and restart of the database (if the db is restarted the value is set to  the current max+1). The owner of the token issued before could now access all data and function of the new account. Therefore the autoincrement in accounts was replaced by a mechanism that saves the last issued id - even if it doesnt exist anymore.